🚚 Move discover button to the all events page
